Added new API callback handlers for aliases, transactions and users

+++ matebot_telegram/features/api_callbacks/user.py
@@ -1,28 +1,30 @@
 """
-API callback handler collection for internal purposes only
+API callback handlers for the events USER_SOFTLY_DELETED and USER_UPDATED
 """
-
-import logging
 
 from matebot_sdk import schemas
 
 from .. import _app
-from ... import persistence
+from ... import models
 
 
 @_app.dispatcher.register_for(schemas.EventType.USER_SOFTLY_DELETED)
-async def _handle_user_softly_deleted(event: schemas.Event):
+async def handle_user_softly_deleted(event: schemas.Event):
     user_id = int(event.data["id"])
     user = (await _app.client.get_users(id=user_id))[0]
     assert not user.active
 
     with _app.client.get_new_session() as session:
-        user = session.query(persistence.TelegramUser).filter_by(user_id=user.id).get()
-        telegram_id = user.telegram_id
-        popped_messages = _app.client.shared_messages.pop_all_messages_by_chat(chat_id=telegram_id)
-        logging.getLogger("api-callback").info(f"Deleting telegram user {telegram_id} (core user {user_id}) ...")
-        session.delete(user)
-        session.query(persistence.RegistrationProcess).filter_by()
+        user = session.query(models.TelegramUser).filter_by(user_id=user.id).get()
+        if user is not None:
+            telegram_id = user.telegram_id
+            popped_messages = _app.client.shared_messages.pop_all_messages_by_chat(chat_id=telegram_id)
+            _app.event_logger.info(f"Deleting telegram user {telegram_id} (core user {user_id}) ...")
+            session.delete(user)
+        registrations = session.query(models.RegistrationProcess).filter_by(telegram_id=telegram_id).all()
+        for r in registrations:
+            _app.event_logger.debug(f"Dropping registration process {r.id} (created: {r.created})")
+            session.delete(r)
         session.commit()
 
     text = (
@@ -38,3 +40,9 @@ async def _handle_user_softly_deleted(event: schemas.Event):
             msg.chat_id,
             msg.message_id
         )
+
+
+@_app.dispatcher.register_for(schemas.EventType.USER_UPDATED)
+async def handle_user_softly_deleted(event: schemas.Event):
+    # TODO: Implement this in case PTB's `user_data` is used purposefully; currently unnecessary function
+    pass
